[file organization] Rename folder names to allow markers to work again
pytest has an issue dealing with sub-folders with the same name and being
included from. In this case, the problem was with several components has
'args' as sub-folder to hold argument parsing utilities.

This issue was initially caught when we were tryign to run the whole test
suite. Attemp was made to address this issue. However there are more in
pytest that cannot tolerate same name sub-folders.

Change 'args' to '<feature>_args' and change conftest.py accordingly.
