+++ tests/arp/conftest.py
@@ -1,4 +1,4 @@
-from args.wr_arp_args import add_wr_arp_args
+from arp_args.wr_arp_args import add_wr_arp_args
 
 # WR-ARP pytest arguments
 def pytest_addoption(parser):
